fix(aoc-dx): run, checking if readme exists before read it

+++ lib/aoc/run.sh
@@ -26,18 +26,22 @@ ensure_input_file_exists() {
 try_to_extract_answers_from_website() {
     curl -s -b session=$(cat session.cookie) https://adventofcode.com/$year/day/$(echo $day | sed 's/^0*//') -o $year/day$day/_readme1.html
 
-    answers=($(grep -oP '<p>Your puzzle answer was <code>\K[^<]+' $year/day$day/_readme1.html))
-    if [ ${#answers[@]} -ne 0 ]; then
-        mkdir -p $year/day$day/data
+    if [ -f "$year/day$day/_readme1.html" ]; then
 
-        printf "%s" "${answers[0]}" > $year/day$day/data/output.part1.txt
-        if [ ${#answers[@]} -eq 2 ]; then
-            printf "%s" "${answers[1]}" > $year/day$day/data/output.part2.txt
+        answers=($(grep -oP '<p>Your puzzle answer was <code>\K[^<]+' $year/day$day/_readme1.html))
+        if [ ${#answers[@]} -ne 0 ]; then
+            mkdir -p $year/day$day/data
+
+            printf "%s" "${answers[0]}" > $year/day$day/data/output.part1.txt
+            if [ ${#answers[@]} -eq 2 ]; then
+                printf "%s" "${answers[1]}" > $year/day$day/data/output.part2.txt
+            fi
+
         fi
 
-    fi
+        rm -rf $year/day$day/_readme*.html
 
-    rm $year/day$day/_readme*.html
+    fi
 }
 
 run_watch_mode() {
